Assign Quality Tag to a Segment Based on maf GMM Metrics and coverage variance

Assigns a quality tag ("PASS" or "FAILED") to a segment based on the maf GMM weight and probe count thresholds.

Usage

AddQualTag(
  Chromosome,
  MAF,
  MAF_gmm_weight,
  MAF_Probes,
  MAF_gmm_G,
  snpmin,
  balance_tag,
  cov_mad,
  sampletype
)

Arguments

Chromosome

Character.

MAF

Numeric. Estimated MAF value.

MAF_gmm_weight

Numeric. The mixture weight of the most prominent maf GMM cluster.

MAF_Probes

Integer. The number of nonzero maf probes in the segment.

MAF_gmm_G

Integer. Number of maf GMM clusters.

snpmin

Integer. The minimum nonzero SNP count.

balance_tag

Character. Balanced test result of BAF of the segment.

cov_mad

Numeric. MAD of coverage.

sampletype

Character. ffpe or ff.

Value

Character. "PASS" if the segment passes quality checks, "FAILED" otherwise.

Details

A segment is marked as "PASS" if the maf GMM weight is at least 0.5, the probe count is at least twice the minimum SNP count (snpmin), coverage MAD is less than 2. Otherwise marked as "FAILED". If the maf GMM weight is 0.5 to 0.8, the balance tag is balanced, but the MAF is < 0.48, it is also FAILED.

Examples

AddQualTag(Chromosome = "1", MAF_gmm_weight = 0.4, MAF_Probes = 20, MAF_gmm_G = 2, snpmin = 7, balance_tag = "balanced", cov_mad = 1, sampletype="ffpe")
#> Error in AddQualTag(Chromosome = "1", MAF_gmm_weight = 0.4, MAF_Probes = 20,     MAF_gmm_G = 2, snpmin = 7, balance_tag = "balanced", cov_mad = 1,     sampletype = "ffpe"): argument "MAF" is missing, with no default
AddQualTag(Chromosome = "X", MAF_gmm_weight = 0.2, MAF_Probes = 10, MAF_gmm_G = 1, snpmin = 7, balance_tag = "balanced", cov_mad = 1, sampletype="ff")
#> Error in AddQualTag(Chromosome = "X", MAF_gmm_weight = 0.2, MAF_Probes = 10,     MAF_gmm_G = 1, snpmin = 7, balance_tag = "balanced", cov_mad = 1,     sampletype = "ff"): argument "MAF" is missing, with no default