Run AI Segmentation Workflow

This function executes the full AI segmentation workflow, including reading input files, merging segments, BAF processing, and writing results.

Usage

RunAIsegmentation(
  seg,
  cov,
  ai,
  gender,
  out_dir,
  prefix,
  ai_pon,
  mergeai = 0.15,
  mergecov = 0.2,
  snpmin = 7,
  minsnpcov = 20,
  minsnpcallaicutoff = 10,
  mergecovminsize = 5e+05,
  maxgap = 2e+06,
  snpnum = 30,
  maxbinsize = 5e+06,
  minbinsize = 5e+05,
  aitype = "gatk",
  segmethod = "cbs",
  cbssmooth = "no"
)

Arguments

seg

Path to the GATK segment file.

cov

Path to the GATK denoised coverage count file.

ai

Path to the BAF file or allelic count file.

gender

Character, "male" or "female".

out_dir

Output directory path.

prefix

Output file prefix.

ai_pon

Path to PON Rdata. AI panel of normal generated by PONAIprocess.

mergeai

Numeric, MAF difference threshold for merging segments (default: 0.15).

mergecov

Numeric, CNV difference threshold for merging segments (default: 0.2).

snpmin

Numeric, minimum SNPs for MAF segmentation (default: 7).

minsnpcov

Numeric, minimum coverage of SNPs to included (default: 20).

minsnpcallaicutoff

Numeric, minimum SNPs for reliable CNLOH/GAINLOH (default: 10).

mergecovminsize

Numeric, minimum size for GATK segment merge (default: 500000).

maxgap

Maximum gap size inside a bin. If exceed then start another bin.( default: 1000000)

snpnum

SNP number in each bin.( default: 30 )

maxbinsize

Maximum bin size.( default: 5000000 )

minbinsize

Minimum bin size.( default: 500000 )

aitype

Character. Type of allelic imbalance data: "gatk" or "other" or "dragen" If "GATK", the input must include columns CONTIG, POSITION, ALT_COUNT, REF_COUNT, REF_NUCLEOTIDE, and ALT_NUCLEOTIDE. If "dragen", the input must include columns contig, start, stop, refAllele, allele1, allele2, allele1Count, allele2Count, allele1AF, and allele2AF. If "other", the input must include columns CONTIG, POSITION, ALT_COUNT, REF_COUNT, REF_NUCLEOTIDE, and ALT_NUCLEOTIDE.

segmethod

Character. Segmentation method to use: if "merge", perform stepwise merging; if "cbs", perform CBS (circular binary segmentation).

cbssmooth

Character. If using the "cbs" segmentation method, set to "yes" to apply smoothing before segmentation, or "no" to skip smoothing.

Value

Invisibly returns the output file path.