Second Refinement of Segment Calls for Coverage/AI Profile Mismatches

Further refines segment calls where the coverage profile and allelic imbalance (AI) profile do not match, by considering alternative copy number states and updating calls and expected values.

Usage

RefineCallsSecond(df, results, final_mu, final_rho, gender, callcov)

Arguments

df

Data frame or tibble. Refined segment calls (see RefineCalls), must include columns as in col_name.

results

Data frame or tibble. All likelihood results for possible copy number states, must include columns rho, mu, index, CN, etc.

final_mu

Numeric. The selected value of scale factor (mu).

final_rho

Numeric. The selected value of tumor purity (rho).

gender

Character. Sample gender ("male" or "female").

callcov

Numeric. Subclonal events calling cutoff based on coverage.

Value

A data frame or tibble with further refined segment calls, including updated columns such as ccf_COV, ccf_final, and CN_mix.

Details

Segments with discordant coverage and AI profiles are re-evaluated by searching for alternative copy number states in the likelihood results, and calls are updated accordingly.

Examples

# RefineCallsSecond(df, results, final_mu = 1, final_rho = 0.7, gender = "female", callcov = 0.3)